Update 2
Fix .gitignore

Restructure injected commands.

Added debugMode to globalInfo.

Implemented custom storage solution for user scripts.

Added a backup mechanism to inject scripts safely. When a page is
preloaded, injected extension scripts are subject to CSP. The work
around is to send the script to the global page and verify its safety.

Renamed locationEnabled to forLocation for consistency.

Additional checks for a proper function header in deepInject.

Bug fixes.