Listen
PageNotification style improvements.

Fixed showPopover command.

Notifications created by user scripts will be sub-titled with the user script name.

Added the ability to create high-priority notifications. These are not affected by "close all" and remain at the top.

Improved the reliability of notifications that should only display once.

Scroll on the top of a notification to either spread the stack or bring it back together.

Alert notifications will now always show the page URL.

Resolves an issue with some notifications not stacking.

Enabling/disabling a kind will now properly update globalSetting and enabled specials.

A notification is displayed when a non-developer mode script is updated.

Bringing forward a notification is now guaranteed to be in front.

The top most notification will now always be visible, i.e., not under a stack.