Phlox - Alpha 10
Added the ability to turn easy lists on/off and to add custom lists.

Added button support to settings. This adds the update now button for easy lists and delete snapshots.

Added fill support for descriptions. This allows the last update time to be shown for easy lists.

Added addition validation support to settings.

Improved rule list is more readable.

Always exclude user rules from Rule#forLocation.

If switching to the Rules tab and the easy list tab was active, change it to temporary before switching.